+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Browser's sessionStorage is not used anymore for user data
 - Relocate 'datasourceID' property in layers and define explicit schema for
   layers, reports and dashboards
+- Configuration object for 'mailer' key has changed, check config/default.js
 
 ### Fixed